форма для надаеления пользователя правами продавца рабочий релиз v.1 завтра будет поиск багов

+++ app/Http/Controllers/Admin/ProductController.php
@@ -65,10 +65,13 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, Product $product)
     {
-        Storage::delete($product->image);
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+
+        $params['image'] =  $product->image;
+        if ($request->file('image') != null) {
+            Storage::delete($product->image);
+            $params['image'] = $request->file('image')->store('products');
+        }
 
         foreach (['new', 'hit', 'recommend'] as $fieldName) {
             if (!isset($params[$fieldName])) {
